apply revert from main repo
