+++ src/msbuild/src/MSBuild.UnitTests/CommandLineSwitches_Tests.cs
@@ -622,8 +622,7 @@ namespace Microsoft.Build.UnitTests
         public void TargetsSwitchParameter()
         {
             CommandLineSwitches switches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
-            parser.GatherCommandLineSwitches(new List<string>() { "/targets:targets.txt" }, switches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>() { "/targets:targets.txt" }, switches);
 
             switches.HaveErrors().ShouldBeFalse();
             switches[CommandLineSwitches.ParameterizedSwitch.Targets].ShouldBe(new[] { "targets.txt" });
@@ -633,8 +632,7 @@ namespace Microsoft.Build.UnitTests
         public void TargetsSwitchDoesNotSupportMultipleOccurrences()
         {
             CommandLineSwitches switches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
-            parser.GatherCommandLineSwitches(new List<string>() { "/targets /targets" }, switches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>() { "/targets /targets" }, switches);
 
             switches.HaveErrors().ShouldBeTrue();
         }
@@ -711,9 +709,8 @@ namespace Microsoft.Build.UnitTests
         public void GraphBuildSwitchCanHaveParameters()
         {
             CommandLineSwitches switches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string> { "/graph", "/graph:true;  NoBuild  ;;  ;", "/graph:foo" }, switches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string> { "/graph", "/graph:true;  NoBuild  ;;  ;", "/graph:foo" }, switches);
 
             switches[CommandLineSwitches.ParameterizedSwitch.GraphBuild].ShouldBe(new[] { "true", "  NoBuild  ", "  ", "foo" });
 
@@ -724,9 +721,8 @@ namespace Microsoft.Build.UnitTests
         public void GraphBuildSwitchCanBeParameterless()
         {
             CommandLineSwitches switches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string> { "/graph" }, switches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string> { "/graph" }, switches);
 
             switches[CommandLineSwitches.ParameterizedSwitch.GraphBuild].ShouldBe(Array.Empty<string>());
 
@@ -737,9 +733,8 @@ namespace Microsoft.Build.UnitTests
         public void InputResultsCachesSupportsMultipleOccurrence()
         {
             CommandLineSwitches switches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>() { "/irc", "/irc:a;b", "/irc:c;d" }, switches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>() { "/irc", "/irc:a;b", "/irc:c;d" }, switches);
 
             switches[CommandLineSwitches.ParameterizedSwitch.InputResultsCaches].ShouldBe(new[] { null, "a", "b", "c", "d" });
 
@@ -750,9 +745,8 @@ namespace Microsoft.Build.UnitTests
         public void OutputResultsCache()
         {
             CommandLineSwitches switches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>() { "/orc:a" }, switches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>() { "/orc:a" }, switches);
 
             switches[CommandLineSwitches.ParameterizedSwitch.OutputResultsCache].ShouldBe(new[] { "a" });
 
@@ -763,9 +757,8 @@ namespace Microsoft.Build.UnitTests
         public void OutputResultsCachesDoesNotSupportMultipleOccurrences()
         {
             CommandLineSwitches switches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>() { "/orc:a", "/orc:b" }, switches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>() { "/orc:a", "/orc:b" }, switches);
 
             switches.HaveErrors().ShouldBeTrue();
         }
@@ -1295,9 +1288,8 @@ namespace Microsoft.Build.UnitTests
         public void ProcessWarnAsErrorSwitchNotSpecified()
         {
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>(new[] { "" }), commandLineSwitches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[] { "" }), commandLineSwitches);
 
             Assert.Null(MSBuildApp.ProcessWarnAsErrorSwitch(commandLineSwitches));
         }
@@ -1311,9 +1303,8 @@ namespace Microsoft.Build.UnitTests
             ISet<string> expectedWarningsAsErrors = new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "a", "B", "c", "D", "e" };
 
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>(new[]
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[]
             {
                 "\"/warnaserror: a,B ; c \"", // Leading, trailing, leading and trailing whitespace
                 "/warnaserror:A,b,C",         // Repeats of different case
@@ -1337,9 +1328,8 @@ namespace Microsoft.Build.UnitTests
         public void ProcessWarnAsErrorSwitchEmptySwitchClearsSet()
         {
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>(new[]
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[]
             {
                 "/warnaserror:a;b;c",
                 "/warnaserror",
@@ -1361,9 +1351,8 @@ namespace Microsoft.Build.UnitTests
             ISet<string> expectedWarningsAsErors = new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "e", "f", "g" };
 
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>(new[]
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[]
             {
                 "/warnaserror:a;b;c",
                 "/warnaserror",
@@ -1384,9 +1373,8 @@ namespace Microsoft.Build.UnitTests
         public void ProcessWarnAsErrorSwitchEmpty()
         {
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>(new[] { "/warnaserror" }), commandLineSwitches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[] { "/warnaserror" }), commandLineSwitches);
 
             ISet<string> actualWarningsAsErrors = MSBuildApp.ProcessWarnAsErrorSwitch(commandLineSwitches);
 
@@ -1402,11 +1390,10 @@ namespace Microsoft.Build.UnitTests
         public void ProcessWarnAsMessageSwitchEmpty()
         {
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
             // Set "expanded" content to match the placeholder so the verify can use the exact resource string as "expected."
             string command = "{0}";
-            parser.GatherCommandLineSwitches(new List<string>(new[] { "/warnasmessage" }), commandLineSwitches, command);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[] { "/warnasmessage" }), commandLineSwitches, command);
 
             VerifySwitchError(commandLineSwitches, "/warnasmessage", AssemblyResources.GetString("MissingWarnAsMessageParameterError"));
         }
@@ -1423,15 +1410,13 @@ namespace Microsoft.Build.UnitTests
                 env.SetEnvironmentVariable("ENVIRONMENTVARIABLE", string.Empty);
 
                 CommandLineSwitches commandLineSwitches = new();
-                CommandLineParser parser = new CommandLineParser();
-
                 string fullCommandLine = "msbuild validProject.csproj %ENVIRONMENTVARIABLE%";
-                parser.GatherCommandLineSwitches(new List<string>() { "validProject.csproj", "%ENVIRONMENTVARIABLE%" }, commandLineSwitches, fullCommandLine);
+                MSBuildApp.GatherCommandLineSwitches(new List<string>() { "validProject.csproj", "%ENVIRONMENTVARIABLE%" }, commandLineSwitches, fullCommandLine);
                 VerifySwitchError(commandLineSwitches, "%ENVIRONMENTVARIABLE%", String.Format(AssemblyResources.GetString("EnvironmentVariableAsSwitch"), fullCommandLine));
 
                 commandLineSwitches = new();
                 fullCommandLine = "msbuild %ENVIRONMENTVARIABLE% validProject.csproj";
-                parser.GatherCommandLineSwitches(new List<string>() { "%ENVIRONMENTVARIABLE%", "validProject.csproj" }, commandLineSwitches, fullCommandLine);
+                MSBuildApp.GatherCommandLineSwitches(new List<string>() { "%ENVIRONMENTVARIABLE%", "validProject.csproj" }, commandLineSwitches, fullCommandLine);
                 VerifySwitchError(commandLineSwitches, "%ENVIRONMENTVARIABLE%", String.Format(AssemblyResources.GetString("EnvironmentVariableAsSwitch"), fullCommandLine));
             }
         }
@@ -1445,9 +1430,8 @@ namespace Microsoft.Build.UnitTests
             ISet<string> expectedWarningsAsMessages = new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "a", "B", "c", "D", "e" };
 
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>(new[]
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[]
             {
                 "\"/warnasmessage: a,B ; c \"", // Leading, trailing, leading and trailing whitespace
                 "/warnasmessage:A,b,C",         // Repeats of different case
@@ -1471,9 +1455,8 @@ namespace Microsoft.Build.UnitTests
         public void ProcessProfileEvaluationEmpty()
         {
             CommandLineSwitches commandLineSwitches = new CommandLineSwitches();
-            CommandLineParser parser = new CommandLineParser();
 
-            parser.GatherCommandLineSwitches(new List<string>(new[] { "/profileevaluation" }), commandLineSwitches);
+            MSBuildApp.GatherCommandLineSwitches(new List<string>(new[] { "/profileevaluation" }), commandLineSwitches);
             commandLineSwitches[CommandLineSwitches.ParameterizedSwitch.ProfileEvaluation][0].ShouldBe("no-file");
         }
 
@@ -1565,7 +1548,11 @@ namespace Microsoft.Build.UnitTests
             using TestEnvironment testEnvironment = TestEnvironment.Create();
             string project = testEnvironment.CreateTestProjectWithFiles("project.proj", projectContent).ProjectFile;
 
+#if FEATURE_GET_COMMANDLINE
             MSBuildApp.Execute(@"msbuild.exe " + project + " /t:foo.bar").ShouldBe(MSBuildApp.ExitType.SwitchError);
+#else
+            MSBuildApp.Execute(new[] { @"msbuild.exe", project, "/t:foo.bar" }).ShouldBe(MSBuildApp.ExitType.SwitchError);
+#endif
         }
 
         /// <summary>
