+++ src/msbuild/src/MSBuild/CommandLineSwitches.cs
